+++ docs/hgvs_railroad.rst
@@ -3,8 +3,8 @@ HGVS Railroad Diagram
 
   | Generated from `hgvs (https://github.com/biocommons/hgvs) <https://github.com/biocommons/hgvs>`_
   | 1b1f788ef473+ default tip
-  | See the source code for the grammar itself: ``src/hgvs/_data/hgvs.pymeta``
-    (OMeta/parsley) and ``src/hgvs/grammar.py`` (pyparsing).
+  | See the source code for the grammar itself: ``src/hgvs/parsers/_data/hgvs.pymeta``
+    (OMeta/parsley) and ``src/hgvs/parsers/pyparsing_grammar.py`` (pyparsing).
 
 .. note:: These diagrams are not regenerated automatically and may lag the
    grammar.
@@ -271,5 +271,3 @@ Remaining rules
   :align: center
 .. image:: hgvs_railroad/term3.svg
   :align: center
-
-
