+++ src/hgvs/parsers/__init__.py
@@ -1,0 +1,3 @@
+from hgvs.parsers.base import OMETA_GRAMMAR, PYPARSING_GRAMMAR, Parser
+
+__all__ = ["OMETA_GRAMMAR", "PYPARSING_GRAMMAR", "Parser"]
