+++ tests/issues/test_730.py
@@ -2,10 +2,10 @@ import os
 import unittest
 
 import pytest
-from support import CACHE
 
 import hgvs.assemblymapper
 import hgvs.dataproviders.uta
+from support import CACHE
 
 
 @pytest.mark.issues
@@ -23,14 +23,14 @@ class Test_Issues(unittest.TestCase):
         self.am37 = hgvs.assemblymapper.AssemblyMapper(
             self.hdp, replace_reference=True, assembly_name="GRCh37", alt_aln_method="splign"
         )
-        self.hp = hgvs.parser.Parser()
+        self.hp = hgvs.parsers.Parser()
 
     def test_730_format_startloss_as_configured(self):
         """
         Parse a start loss and make sure the one letter and three letter protein genotypes come back as expected.
         """
         var_g = self.hp.parse_hgvs_variant("NC_000016.9:g.89985662_89985667del")
-        var_c = self.am37.g_to_c(var_g, str("NM_002386.3"))
+        var_c = self.am37.g_to_c(var_g, "NM_002386.3")
         var_p = self.am37.c_to_p(var_c)
         var_p_one_letter = var_p.format(conf={"p_3_letter": False})
         var_p_three_letter = var_p.format(conf={"p_3_letter": True})
