+++ tests/support/crosschecker.py
@@ -2,7 +2,7 @@ import itertools
 
 import hgvs.dataproviders.uta
 import hgvs.edit
-import hgvs.parser
+import hgvs.parsers
 import hgvs.sequencevariant
 import hgvs.variantmapper
 from hgvs.exceptions import HGVSDataNotAvailableError
@@ -40,14 +40,14 @@ class CrossChecker:
 
         variants = sorted(variants, key=lambda v: v.type)
         binned_variants = {k: [] for k in "cgmnrp"}
-        binned_variants.update(
-            {g: list(gi) for g, gi in itertools.groupby(variants, lambda v: v.type)}
-        )
+        binned_variants.update({
+            g: list(gi) for g, gi in itertools.groupby(variants, lambda v: v.type)
+        })
         binned_variants["t"] = binned_variants["c"] + binned_variants["n"]
 
-        assert len(binned_variants["g"]) == len(
-            [v.ac for v in binned_variants["g"]]
-        ), "variants have multiple alignments"
+        assert len(binned_variants["g"]) == len([v.ac for v in binned_variants["g"]]), (
+            "variants have multiple alignments"
+        )
 
         # g -> t: for each g., map to each transcript accession.
         for g_var in binned_variants["g"]:
@@ -57,9 +57,7 @@ class CrossChecker:
                 except HGVSDataNotAvailableError:
                     continue
                 if t_var != r:
-                    return "g_to_t({g_var},{t_var.ac}): got {r}; expected {t_var}".format(
-                        g_var=g_var, t_var=t_var, r=r
-                    )
+                    return f"g_to_t({g_var},{t_var.ac}): got {r}; expected {t_var}"
 
         # t -> g: for each t., map to each genomic accession
         for t_var in binned_variants["t"]:
@@ -69,9 +67,7 @@ class CrossChecker:
                 except HGVSDataNotAvailableError:
                     continue
                 if g_var != r:
-                    return "t_to_g({t_var},{g_var.ac}): got {r}; expected {g_var}".format(
-                        g_var=g_var, t_var=t_var, r=r
-                    )
+                    return f"t_to_g({t_var},{g_var.ac}): got {r}; expected {g_var}"
 
         # c -> p: for each c., map to a protein variant and check whether it's in result set
