+++ tests/test_hgvs_grammar.py
@@ -1,11 +1,10 @@
-# -*- coding: utf-8 -*-
 import unittest
 
 import pytest
 
-import hgvs.parser
+import hgvs.parsers
 
-# we're not testing hgvs.parser, but rather merely using it to load the
+# we're not testing hgvs.parsers, but rather merely using it to load the
 # grammar.  See test_hgvs_parser.py for the parser tests
 
 
@@ -14,7 +13,7 @@ class Test_Parser(unittest.TestCase):
     longMessage = True
 
     def setUp(self):
-        self.p = hgvs.parser.Parser(expose_all_rules=True)
+        self.p = hgvs.parsers.Parser(expose_all_rules=True)
         self.grammar = self.p._grammar
 
     def test_parser_basic(self):
