+++ tests/test_hgvs_grammar_equivalence.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Tests that the pyparsing and OMeta/parsley backends behave identically.
 
 The pyparsing grammar is intended to replace OMeta/parsley (issue #505). While
@@ -15,16 +14,16 @@ import unittest
 
 import pyparsing as pp
 
-import hgvs.parser
+import hgvs.parsers
 from hgvs.exceptions import HGVSParseError
 
-BACKENDS = (hgvs.parser.OMETA_GRAMMAR, hgvs.parser.PYPARSING_GRAMMAR)
+BACKENDS = (hgvs.parsers.OMETA_GRAMMAR, hgvs.parsers.PYPARSING_GRAMMAR)
 
 _GAUNTLET_FN = os.path.join(os.path.dirname(__file__), "data", "gauntlet")
 
 
 def _read_gauntlet():
-    with open(_GAUNTLET_FN, "r") as f:
+    with open(_GAUNTLET_FN) as f:
         return [line.strip() for line in f if line.strip() and not line.strip().startswith("#")]
 
 
@@ -39,8 +38,8 @@ def _parse_or_error(parser, variant):
 class TestBackendEquivalence(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.ometa = hgvs.parser.Parser(grammar_fn=hgvs.parser.OMETA_GRAMMAR)
-        cls.pyparsing = hgvs.parser.Parser(grammar_fn=hgvs.parser.PYPARSING_GRAMMAR)
+        cls.ometa = hgvs.parsers.Parser(grammar_fn=hgvs.parsers.OMETA_GRAMMAR)
+        cls.pyparsing = hgvs.parsers.Parser(grammar_fn=hgvs.parsers.PYPARSING_GRAMMAR)
 
     def test_backends_agree_on_gauntlet(self):
         """both backends must return identical results for every gauntlet variant"""
@@ -53,7 +52,7 @@ class TestBackendEquivalence(unittest.TestCase):
                 self.assertEqual(
                     _parse_or_error(self.ometa, variant),
                     _parse_or_error(self.pyparsing, variant),
-                    "backends disagree on {!r}".format(variant),
+                    f"backends disagree on {variant!r}",
                 )
 
     def test_exposed_rules_match(self):
@@ -69,16 +68,16 @@ class TestNoGlobalPyparsingState(unittest.TestCase):
     """pyparsing's whitespace/packrat settings are process-global.
 
     hgvs must not leak them: doing so silently changes the behavior of any
-    other pyparsing grammar in the same process. cf. hgvs.grammar.
+    other pyparsing grammar in the same process. cf. hgvs.parsers.pyparsing_grammar.
     """
 
     def test_building_grammar_does_not_leak_whitespace_default(self):
         before = pp.ParserElement.DEFAULT_WHITE_CHARS
-        hgvs.parser.Parser(grammar_fn=hgvs.parser.PYPARSING_GRAMMAR)
+        hgvs.parsers.Parser(grammar_fn=hgvs.parsers.PYPARSING_GRAMMAR)
         self.assertEqual(pp.ParserElement.DEFAULT_WHITE_CHARS, before)
 
     def test_third_party_grammar_unaffected(self):
-        hgvs.parser.Parser(grammar_fn=hgvs.parser.PYPARSING_GRAMMAR)
+        hgvs.parsers.Parser(grammar_fn=hgvs.parsers.PYPARSING_GRAMMAR)
         expr = pp.Word(pp.alphas) + pp.Word(pp.nums)
         self.assertEqual(expr.parse_string("abc 123").as_list(), ["abc", "123"])
 
@@ -88,7 +87,7 @@ class TestNoGlobalPyparsingState(unittest.TestCase):
         prev = pp.ParserElement.DEFAULT_WHITE_CHARS
         try:
             pp.ParserElement.set_default_whitespace_chars(" \t\n")
-            parser = hgvs.parser.Parser(grammar_fn=hgvs.parser.PYPARSING_GRAMMAR)
+            parser = hgvs.parsers.Parser(grammar_fn=hgvs.parsers.PYPARSING_GRAMMAR)
             for variant in (
                 "NM_01234.5 :c.22+1A>T",
                 "NM_01234.5: c.22+1A>T",
@@ -112,8 +111,8 @@ class TestWhitespaceHandling(unittest.TestCase):
     VARIANT = "NM_01234.5:c.22+1A>T"
 
     def test_ometa_accepts_leading_space_pyparsing_rejects(self):
-        ometa = hgvs.parser.Parser(grammar_fn=hgvs.parser.OMETA_GRAMMAR)
-        pyparsing = hgvs.parser.Parser(grammar_fn=hgvs.parser.PYPARSING_GRAMMAR)
+        ometa = hgvs.parsers.Parser(grammar_fn=hgvs.parsers.OMETA_GRAMMAR)
+        pyparsing = hgvs.parsers.Parser(grammar_fn=hgvs.parsers.PYPARSING_GRAMMAR)
         for padded in (" " + self.VARIANT, "\t" + self.VARIANT, "  " + self.VARIANT):
             with self.subTest(variant=padded):
                 self.assertEqual(str(ometa.parse_hgvs_variant(padded)), self.VARIANT)
@@ -122,7 +121,7 @@ class TestWhitespaceHandling(unittest.TestCase):
 
     def test_both_backends_reject_trailing_whitespace(self):
         for backend in BACKENDS:
-            parser = hgvs.parser.Parser(grammar_fn=backend)
+            parser = hgvs.parsers.Parser(grammar_fn=backend)
             for padded in (self.VARIANT + " ", self.VARIANT + "\t", "\n" + self.VARIANT):
                 with self.subTest(backend=backend, variant=padded):
                     with self.assertRaises(HGVSParseError):
