+++ tests/test_hgvs_variantmapper_cp_altseqbuilder.py
@@ -1,11 +1,9 @@
-# -*- coding: utf-8 -*-
 import os
 import unittest
 
-import support.mock_input_source as mock_input_data_source
-
-import hgvs.parser
+import hgvs.parsers
 import hgvs.utils.altseqbuilder as altseqbuilder
+import support.mock_input_source as mock_input_data_source
 from hgvs.utils.reftranscriptdata import RefTranscriptData
 
 
@@ -13,7 +11,7 @@ class TestAltSeqBuilder(unittest.TestCase):
     # root sequence = ""
     fn = os.path.join(os.path.dirname(__file__), "data", "sanity_cp.tsv")
     _datasource = mock_input_data_source.MockInputSource(fn)
-    _parser = hgvs.parser.Parser()
+    _parser = hgvs.parsers.Parser()
 
     def test_substitution_start(self):
         hgvsc = "NM_999999.1:c.1A>T"
@@ -122,7 +120,7 @@ class TestAltSeqBuilder(unittest.TestCase):
         builder = altseqbuilder.AltSeqBuilder(var, transcript_data)
         insert_result = builder.build_altseq()
         actual_sequence = insert_result[0].transcript_sequence
-        msg = "expected: {}\nactual  : {}".format(expected_sequence, actual_sequence)
+        msg = f"expected: {expected_sequence}\nactual  : {actual_sequence}"
         self.assertEqual(expected_sequence, actual_sequence, msg)
